requirements.txt
together==1.3.1
requests==2.32.3
python-dotenv==1.0.1
flask==2.2.5 
werkzeug==2.2.3 
connexion==2.14.2